Reorganize src into core/models/legacy with compatibility shims
