+++ src/ml_wireless_classification/base/__init__.py
@@ -1,1 +1,8 @@
 # base
+"""Backward-compatible base module exports.
+
+The canonical maintained code now lives under:
+- ``ml_wireless_classification.core``
+- ``ml_wireless_classification.models``
+- ``ml_wireless_classification.legacy``
+"""
