+++ src/ml_wireless_classification/legacy/__init__.py
@@ -1,0 +1,1 @@
+"""Legacy and experimental utilities retained for archive notebooks."""
