+++ src/ml_wireless_classification/models/__init__.py
@@ -1,0 +1,5 @@
+"""Maintained model architectures."""
+
+from ml_wireless_classification.models.rnn_lstm_with_snr import ModulationLSTMClassifier
+
+__all__ = ["ModulationLSTMClassifier"]
