+++ src/ml_wireless_classification/rnn_lstm_w_SNR_ensemble.py
@@ -1,0 +1,10 @@
+"""Legacy compatibility shim.
+
+Archive notebooks historically imported `ModulationLSTMClassifier` from this module.
+The maintained implementation now lives in
+`ml_wireless_classification.models.rnn_lstm_with_snr`.
+"""
+
+from ml_wireless_classification.models.rnn_lstm_with_snr import ModulationLSTMClassifier
+
+__all__ = ["ModulationLSTMClassifier"]
